+++ NEWS.md
@@ -1,4 +1,7 @@
 
+# renv (development version)
+
+
 # renv 1.1.6
 
 * A new setting, `snapshot.dev`, has been added to control whether development